Add Z-axis support, fix issue with editor undo

+++ emit.c
@@ -31,7 +31,7 @@ void emit_set_rad(emit_desc* e, double rad, double urad) {
     e->urad = urad;
 }
 
-void emit_set_depth(emit_desc* e, int depth) {
+void emit_set_depth(emit_desc* e, float depth) {
     e->depth = depth;
 }
 
@@ -82,7 +82,7 @@ void emit_set_tag(emit_desc* e, uint64_t tag) {
 
 void dbprintf_emit_desc(emit_desc* e) {
     DBPRINTF("struct emit_desc { n=%d, x=%g, y=%g, ux=%g, uy=%g, "
-             "s=%g us=%g, rad=%g, urad=%g, depth=%d, life=%d, ulife=%d, "
+             "s=%g us=%g, rad=%g, urad=%g, depth=%g, life=%d, ulife=%d, "
              "ds=%g, uds=%g, theta=%g, utheta=%g, r=%f, g=%f, b=%f, "
              "ur=%f, ug=%f, ub=%f, force=%d, limit=%d, blender=%d };",
         e->n, e->x, e->y, e->ux, e->uy, e->s, e->us, e->rad, e->urad,
